samba4: update smb.conf.template to disable guest logons by default
Insecure guest logons are on UpToDate Windows 10 1709 and later disabled by default. The option `map to guest` should be set to Samba default value = `Never`, because only then the Windows will ask to enter the password when accesing the share (when there are no valid user logged-in in Windows). If is the option `map to guest` is set to `Bad User` then Windows Client will not connect (when the user, which is logged-in in Windows is not also configured on OpenWrt) and fail with message, that insecure logons are disabled.

https://learn.microsoft.com/en-us/troubleshoot/windows-server/networking/guest-access-in-smb2-is-disabled-by-default

Typo also corrected.

Signed-off-by: Jiří Janata <jirijanata@hotmail.com>