iperf3: handle ssl support as config option
This handles the SSL support for iperf3 as a configuration option rather
than as a separate package (variant).
This way an issue is fixed where libiperf3 fails to build when the ssl
variant of iperf3 is selected. The build failure is caused by the
'--disable-shared' that is added to CONFIGURE_ARGS when selecting the
ssl variant. This prevents the build from generating the libiperf.so.*.

Besides fixing the issue, this also allows to have shared libiperf3 with
SSL support and avoids the need to statically compile iperf3-ssl.

Signed-off-by: Jonas Jelonek <jelonek.jonas@gmail.com>