ethtool-lua: Initial packaging
The ethtool-lua library is a partial re-implementation of the ethtool.
The goal is to provide the CLI queries and configuration options as a
Lua API.

The reason for staring this library, was the desire for a nice and
efficient way to query DSA switch statistics in the
prometheus-node-exporter-lua on OpenWRT devices. Existing suggestions
around the internet focussed mainly on calling the ethtool CLI program
and parsing the output. This is neither elegant nor efficient, as the
collection time for this implementation was 300% higher on my rtl838x
based switch running OpenWRT.

This package can be easily compiled and packaged for OpenWRT, as it was
initially created as an OpenWRT package.

Signed-off-by: Kevin Jilissen <info@kevinjilissen.nl>